Influence/requirements.txt
line-bot-sdk==2.3.0
openai==0.27.8
urllib3==1.26.7
langchain==0.0.137